test(policy-diff): split test_cli_policy_diff into themed modules (#263 F4)
The single 1221-line tests/test_cli_policy_diff.py concentrated five
distinct behavior areas of `policies diff` into one module, raising
maintenance cost and obscuring behavior ownership (review finding F4).

Split the 43 tests verbatim into five focused siblings — report,
json, exit-code, elements, cross-policy — and lift the shared client
stub and PolicyRevision factories into tests/_policy_diff_helpers.py
(mirroring the existing tests/_openapi shared-helper package). Each
module keeps its own thin `stub` fixture delegating to make_stub, so
no fixture is imported (avoiding an unused-import on the fixture name).
Pure test reorganization — no production code and no test logic change.
